Use swift-argument-parser for notification app
This updates the macOS notification app to use 'swift-argument-parser'
for parsing command line arguments. Should be more robust now.
